+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/modelo/ChecklistModeloDao.java
@@ -1,8 +1,5 @@
 package br.com.zalf.prolog.webservice.frota.checklist.modelo;
 
-
-import br.com.zalf.prolog.webservice.colaborador.model.Empresa;
-import br.com.zalf.prolog.webservice.colaborador.model.Unidade;
 import br.com.zalf.prolog.webservice.commons.imagens.Galeria;
 import br.com.zalf.prolog.webservice.commons.imagens.ImagemProLog;
 import br.com.zalf.prolog.webservice.frota.checklist.model.TipoChecklist;
@@ -14,6 +11,8 @@ import br.com.zalf.prolog.webservice.frota.checklist.modelo.model.realizacao.Mod
 import br.com.zalf.prolog.webservice.frota.checklist.modelo.model.realizacao.ModeloChecklistSelecao;
 import br.com.zalf.prolog.webservice.frota.checklist.modelo.model.visualizacao.ModeloChecklistVisualizacao;
 import br.com.zalf.prolog.webservice.frota.checklist.offline.DadosChecklistOfflineChangedListener;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Empresa;
+import br.com.zalf.prolog.webservice.geral.unidade._model.Unidade;
 import br.com.zalf.prolog.webservice.permissao.pilares.FuncaoProLog;
 import org.jetbrains.annotations.NotNull;