+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ordemservico/model/listagem/OrdemServicoListagem.java
@@ -1,10 +1,10 @@
 package br.com.zalf.prolog.webservice.frota.checklist.ordemservico.model.listagem;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Unidade;
 import br.com.zalf.prolog.webservice.commons.gson.Exclude;
 import br.com.zalf.prolog.webservice.commons.gson.RuntimeTypeAdapterFactory;
 import br.com.zalf.prolog.webservice.frota.checklist.ordemservico.model.StatusOrdemServico;
 import br.com.zalf.prolog.webservice.frota.veiculo.model.Veiculo;
+import br.com.zalf.prolog.webservice.geral.unidade._model.Unidade;
 import org.jetbrains.annotations.NotNull;
 
 import java.time.LocalDateTime;
@@ -12,51 +12,45 @@ import java.time.LocalDateTime;
 /**
  * Esta é a superclasse utilizada para instanciar a Listagem de Ordens de Serviço Abertas ou a
  * Listagem de Ordens de Serviço Fechadas.
- *
+ * <p>
  * Created on 09/11/18
  *
  * @author Luiz Felipe (https://github.com/luizfp)
  */
 public abstract class OrdemServicoListagem {
-    /**
-     * Código da Ordem de Serviço.
-     */
-    private Long codOrdemServico;
-
-    /**
-     * Código da {@link Unidade} a qual a Ordem de Serviço pertence.
-     */
-    private Long codUnidadeOrdemServico;
-
-    /**
-     * Placa do {@link Veiculo} a qual a Ordem de Serviço pertence.
-     */
-    private String placaVeiculo;
-
-    /**
-     * Data e Hora em que a Ordem de Serviço foi aberta.
-     */
-    private LocalDateTime dataHoraAbertura;
-
-    /**
-     * Quantidade de Itens para serem resolvidos.
-     */
-    private int qtdItensPendentes;
-
-    /**
-     * Quantidade de Itens já resolvidos.
-     */
-    private int qtdItensResolvidos;
 
     /**
      * O status atual dessa O.S.
      */
     @NotNull
     private final StatusOrdemServico status;
-
     @NotNull
     @Exclude
     private final String tipo;
+    /**
+     * Código da Ordem de Serviço.
+     */
+    private Long codOrdemServico;
+    /**
+     * Código da {@link Unidade} a qual a Ordem de Serviço pertence.
+     */
+    private Long codUnidadeOrdemServico;
+    /**
+     * Placa do {@link Veiculo} a qual a Ordem de Serviço pertence.
+     */
+    private String placaVeiculo;
+    /**
+     * Data e Hora em que a Ordem de Serviço foi aberta.
+     */
+    private LocalDateTime dataHoraAbertura;
+    /**
+     * Quantidade de Itens para serem resolvidos.
+     */
+    private int qtdItensPendentes;
+    /**
+     * Quantidade de Itens já resolvidos.
+     */
+    private int qtdItensResolvidos;
 
     public OrdemServicoListagem(@NotNull final String tipo,
                                 @NotNull final StatusOrdemServico status) {