+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/_model/Afericao.java
@@ -1,10 +1,10 @@
 package br.com.zalf.prolog.webservice.frota.pneu.afericao._model;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
-import br.com.zalf.prolog.webservice.colaborador.model.Unidade;
 import br.com.zalf.prolog.webservice.commons.gson.Exclude;
 import br.com.zalf.prolog.webservice.commons.gson.RuntimeTypeAdapterFactory;
 import br.com.zalf.prolog.webservice.frota.pneu._model.Pneu;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.geral.unidade._model.Unidade;
 import org.jetbrains.annotations.NotNull;
 
 import java.time.LocalDateTime;
@@ -16,8 +16,11 @@ import java.util.List;
  * @author Luiz Felipe (https://github.com/luizfp)
  */
 public abstract class Afericao {
-    private Long codigo;
 
+    @Exclude
+    @NotNull
+    private final TipoProcessoColetaAfericao tipoProcessoColetaAfericao;
+    private Long codigo;
     /**
      * O código da {@link Unidade} onde a aferição foi realizada.
      */
@@ -28,22 +31,25 @@ public abstract class Afericao {
      * Armazena o tempo que o colaborador levou para realizar a aferição, em milisegundos.
      */
     private long tempoRealizacaoAfericaoInMillis;
-
     private TipoMedicaoColetadaAfericao tipoMedicaoColetadaAfericao;
-
-    @Exclude
-    @NotNull
-    private final TipoProcessoColetaAfericao tipoProcessoColetaAfericao;
 
     public Afericao(@NotNull final TipoProcessoColetaAfericao tipoProcessoColetaAfericao) {
         this.tipoProcessoColetaAfericao = tipoProcessoColetaAfericao;
+    }
+
+    @NotNull
+    public static RuntimeTypeAdapterFactory<Afericao> provideTypeAdapterFactory() {
+        return RuntimeTypeAdapterFactory
+                .of(Afericao.class, "tipoProcessoColetaAfericao")
+                .registerSubtype(AfericaoPlaca.class, TipoProcessoColetaAfericao.PLACA.asString())
+                .registerSubtype(AfericaoAvulsa.class, TipoProcessoColetaAfericao.PNEU_AVULSO.asString());
     }
 
     public Colaborador getColaborador() {
         return colaborador;
     }
 
-    public void setColaborador(Colaborador colaborador) {
+    public void setColaborador(final Colaborador colaborador) {
         this.colaborador = colaborador;
     }
 
@@ -51,7 +57,7 @@ public abstract class Afericao {
         return codigo;
     }
 
-    public void setCodigo(Long codigo) {
+    public void setCodigo(final Long codigo) {
         this.codigo = codigo;
     }
 
@@ -67,7 +73,7 @@ public abstract class Afericao {
         return dataHora;
     }
 
-    public void setDataHora(LocalDateTime dataHora) {
+    public void setDataHora(final LocalDateTime dataHora) {
         this.dataHora = dataHora;
     }
 
@@ -75,7 +81,7 @@ public abstract class Afericao {
         return tempoRealizacaoAfericaoInMillis;
     }
 
-    public void setTempoRealizacaoAfericaoInMillis(long tempoRealizacaoAfericaoInMillis) {
+    public void setTempoRealizacaoAfericaoInMillis(final long tempoRealizacaoAfericaoInMillis) {
         this.tempoRealizacaoAfericaoInMillis = tempoRealizacaoAfericaoInMillis;
     }
 
@@ -83,7 +89,7 @@ public abstract class Afericao {
         return tipoMedicaoColetadaAfericao;
     }
 
-    public void setTipoMedicaoColetadaAfericao(TipoMedicaoColetadaAfericao tipoMedicaoColetadaAfericao) {
+    public void setTipoMedicaoColetadaAfericao(final TipoMedicaoColetadaAfericao tipoMedicaoColetadaAfericao) {
         this.tipoMedicaoColetadaAfericao = tipoMedicaoColetadaAfericao;
     }
 
@@ -94,12 +100,4 @@ public abstract class Afericao {
 
     @NotNull
     public abstract List<Pneu> getPneusAferidos();
-
-    @NotNull
-    public static RuntimeTypeAdapterFactory<Afericao> provideTypeAdapterFactory() {
-        return RuntimeTypeAdapterFactory
-                .of(Afericao.class, "tipoProcessoColetaAfericao")
-                .registerSubtype(AfericaoPlaca.class, TipoProcessoColetaAfericao.PLACA.asString())
-                .registerSubtype(AfericaoAvulsa.class, TipoProcessoColetaAfericao.PNEU_AVULSO.asString());
-    }
 }