+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/pneutiposervico/PneuTipoServicoDao.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.frota.pneu.pneutiposervico;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Empresa;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Empresa;
 import br.com.zalf.prolog.webservice.commons.OrderByClause;
 import br.com.zalf.prolog.webservice.frota.pneu.movimentacao._model.origem.OrigemAnalise;
 import br.com.zalf.prolog.webservice.frota.pneu.pneutiposervico._model.PneuTipoServico;