+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/recapadoras/RecapadoraDao.java
@@ -1,7 +1,7 @@
 package br.com.zalf.prolog.webservice.frota.pneu.recapadoras;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
-import br.com.zalf.prolog.webservice.colaborador.model.Empresa;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Empresa;
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.SQLException;