+++ src/main/java/br/com/zalf/prolog/webservice/gente/controlejornada/ControleJornadaConverter.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.gente.controlejornada;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 import br.com.zalf.prolog.webservice.commons.FonteDataHora;
 import br.com.zalf.prolog.webservice.gente.controlejornada.model.Intervalo;
 import br.com.zalf.prolog.webservice.gente.controlejornada.model.Localizacao;