+++ src/main/java/br/com/zalf/prolog/webservice/gente/controlejornada/ajustes/model/exibicao/MarcacoesDiaColaborador.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.gente.controlejornada.ajustes.model.exibicao;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 import org.jetbrains.annotations.NotNull;
 
 /**