+++ src/main/java/br/com/zalf/prolog/webservice/gente/faleConosco/FaleConosco.java
@@ -1,7 +1,7 @@
 package br.com.zalf.prolog.webservice.gente.faleConosco;
 
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 
 import java.time.LocalDateTime;