+++ src/main/java/br/com/zalf/prolog/webservice/gente/faleConosco/FaleConoscoDaoImpl.java
@@ -2,7 +2,7 @@ package br.com.zalf.prolog.webservice.gente.faleConosco;
 
 import br.com.zalf.prolog.webservice.database.DatabaseConnection;
 import br.com.zalf.prolog.webservice.TimeZoneManager;
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;