+++ src/main/java/br/com/zalf/prolog/webservice/gente/solicitacaoFolga/SolicitacaoFolga.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.gente.solicitacaoFolga;
 
-import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
+import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 
 import java.util.Date;